refactor: update claim page to the same structure
